Update README.md
This package, based on 4sU-seq data and RNA-seq data, analyzes the transcription, processing and degradation rates of genes. pulseTD can not only recognize the transcriptional dynamic rate of the measurement time points, but also obtain continuous changes in transcriptional dynamics. More importantly, it is able to predict the trend of mRNA transcription and expression changes in the future. In terms of performance, pulseTD has better robustness and accuracy than other methods.